\name{Rdsdp::dsdp.readsdpa}
\alias{dsdp.readsdpa}
\alias{readsdpa}
\title{
Solving semidefinite programs reading from SDPA format files.
}
\description{
Function to read the semidefinite program input data in SDPA format and solve it.
}

\usage{
dsdp.readsdpa(sdpa_filename, options_filename="")
}

\arguments{
	\item{sdpa_filename}{The location of the SDPA input file.}
	\item{options_filename}{The location of the OPTIONS file [default ""].}
}




